Rename folder from tests to pytests and update README.md
