+++ .env.example
@@ -1,0 +1,6 @@
+BOT_TOKEN = "Token from BotFather"
+CONFIG_PATH = "configs/"
+DEFAULT_JSON = "${CONFIG_PATH}default_config.json"
+SAVED_JSON = "${CONFIG_PATH}saved_config.json"
+MUSIC_FILE = "music.csv"
+CHAT_ID = 123456789
