+++ pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 addopts = --tb=short --strict
-testpaths = tests
+testpaths = pytests
 markers =
     smoke: smoke tests
     work_with_dj: dj_commands tests
